printf testeado

+++ ft_printf.h
@@ -14,6 +14,7 @@
 # define FT_PRINTF_H
 # include <stdarg.h>
 # include <unistd.h>
+# include <stdio.h>
 
 void	print_char(char c);
 int		ft_printf(char const *input, ...);
@@ -21,7 +22,7 @@ int		print_str(char *str);
 int		print_nbr(int n);
 int		print_nbr_unsigned(unsigned int n);
 void	print_per(void);
-int		print_hex(unsigned int n, int uppercase);
+int		print_hex(unsigned long long n, int uppercase);
 int		print_ptr(unsigned long long ptr);
 
 #endif
